Updated to use getNormalizedRelativePath
